[Refactor] Participant Converter 변경
